+++ torch_points_kernels/__init__.py
@@ -1,4 +1,5 @@
 from .torchpoints import *
 from .knn import knn
+from .points_cpu import GridSampler
 
-__all__ = ["ball_query", "furthest_point_sample", "grouping_operation", "three_interpolate", "three_nn", "knn"]
+__all__ = ["ball_query", "furthest_point_sample", "grouping_operation", "three_interpolate", "three_nn", "knn","GridSampler" ]